Fixed 2 unit test bugs and added choosing of api setting
Fixed first unit test bug by adding missing http dependency to vtmapi.service.spec.ts and fixed races.component.spec.ts unit test bug by setting default api setting to be `dnd` and adding select element to html to select the api setting you want to use. The data for the races component doesn't get retrieved after the component is initialized currently so it is kind of useless for now. But in the future it could be extended so that the data is retrieved whenever the select element value is change or something