Created reactive new entry form modal for newEntry component
Created a new entry form that is reactive and is in a modal for the newEntry component. 
Also added the ability to save a new entry and then it emits an event for the parent component to use the new entry. 
Updated the races component to listen to the emitted event and update the races variable with the new race.
Finally, added unit tests for the added code.