Added settings service and modified existing files to work with it
Added the settings service and settings component so the settings can be change. Also updated the other components to work with the new settings service. This includes modifying the api selector service which all the other components depend on.

Secondly, added the APP_INITIALIZER token to the app module file so that the settings service is run first on app startup so that the initFunction in it is run. This lets the settings service load the settings that will be used by all the other components on the site. Without doing that the other components won't know what api to use if they are the first page opened in the site instead of the settings page.

Finally, added a check in the table-detail component for an undefined url since it would randomly send a data request with an undefined url for some reason. Will need to be looked at later