Updated the unit tests that interacted with the settings service
Updated the unit tests that interacted with the settings service so that the apiSelectorService was spied on and when a call for the api was made it just responded with the appropriate api instead of calling the settings service. This is because in the unit testing the app module doesn't actually initialize the settings service so errors appear with an undefined settings variable in the settings service.

Also fixed the table detail view unit tests because the file was importing a component when it should have been providing the component.

Also fixed the dashboard layout unit test.