Moved services into a common folder
I moved the services into a common folder since they are being utilized throughout the application and not just in the spots where they are defined. This ends up being confusing since they are not just being used where they are located so it makes sense to me to put them all in one spot (a services folder) where they can be imported from.